fixed typo in call to binary_tree_height

+++ 14-binary_tree_balance.c
@@ -1,4 +1,5 @@
 #include "binary_trees.h"
+
 /**
  * binary_tree_balance - Measures the balance factor of a binary tree
  * @tree: Pointer to the root node of the tree to measure the balance factor
@@ -8,14 +9,14 @@
 
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	int left_height;
-	int right_height;
+	size_t left_height;
+	size_t right_height;
 
 	if (tree == NULL)
 	{
 		return (0);
 	}
-	left_height = tree->left ? (int)binary_tree_height(tree->left) : 0;
-	right_height = tree->right ? (int)binary_tree_height(tree->right) : 0;
+	left_height = tree->left ? binary_tree_height(tree->left) : 0;
+	right_height = tree->right ? binary_tree_height(tree->right) : 0;
 	return (left_height - right_height);
 }
